Move exception to logic package
