+++ src/main/java/seedu/duke/logic/command/SaveCommand.java
@@ -1,7 +1,7 @@
 package seedu.duke.logic.command;
 
 
-import seedu.duke.exception.IllegalValueException;
+import seedu.duke.logic.exception.IllegalValueException;
 import seedu.duke.records.RecordList;
 import seedu.duke.records.biometrics.Biometrics;
 import seedu.duke.records.exercise.ExerciseList;
