+++ src/main/java/seedu/duke/logic/exception/IllegalValueException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.duke.logic.exception;
 
 public class IllegalValueException extends DukeException {
     public IllegalValueException(String message) {
