+++ src/main/java/seedu/duke/logic/exception/InvalidDateException.java
@@ -1,4 +1,4 @@
-package seedu.duke.exception;
+package seedu.duke.logic.exception;
 
 public class InvalidDateException extends DukeException {
     public InvalidDateException(String message) {
