+++ src/test/java/seedu/duke/command/RemoveCommandTest.java
@@ -2,7 +2,7 @@ package seedu.duke.command;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import seedu.duke.exception.IllegalValueException;
+import seedu.duke.logic.exception.IllegalValueException;
 import seedu.duke.logic.Parser;
 import seedu.duke.logic.command.Command;
 import seedu.duke.records.RecordList;
