+++ src/test/java/seedu/duke/exercise/ExerciseListTest.java
@@ -1,7 +1,7 @@
 package seedu.duke.exercise;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.exception.IllegalValueException;
+import seedu.duke.logic.exception.IllegalValueException;
 import seedu.duke.logic.Parser;
 import seedu.duke.logic.command.Command;
 import seedu.duke.records.RecordList;
